requirements.txt
pandas
datasets
openai
ragas
langchain-openai
python-dotenv
pydantic